+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+**./src/common/dsktra-tasks.js**
+* New file - Sub functions for Dijkstra pathfinding algorithm.
+
+---
+
 **./src/shortest-route.js**
-* Renamed 'closed' result property to 'closedRoute' (defineDijkstraInfo)
-* Renamed 'closeStatus' variable to 'nodesMatch' (setClosedRoute)
+* Required './common/dsktra-tasks.js'
+* Moved the following functions to 'dsktraTasks'
+	* defineDijkstraInfo
+	* setNodeTable
+	* setClosedRoute
+	* 'sortNodes' as 'sortNodeTable'